requirements.txt
fastapi
uvicorn
pydantic
pytesseract
pdfplumber
pillow
pytest
